feat(web): implemented responsive settings page
