+++ apps/web/src/components/StudentSettings/TopicSelection.tsx
@@ -1,6 +1,6 @@
 import { useFormatMessage } from "@litespace/ui/hooks/intl";
 import { Typography } from "@litespace/ui/Typography";
-import React, { useCallback, useMemo, useState } from "react";
+import React, { useCallback, useEffect, useMemo, useState } from "react";
 import Edit from "@litespace/assets/Edit";
 import { TopicSelectionDialog } from "@litespace/ui/TopicSelectionDialog";
 import { Animate } from "@/components/Common/Animate";
@@ -15,28 +15,42 @@ import { useToast } from "@litespace/ui/Toast";
 import { useUpdateUserTopics } from "@litespace/headless/user";
 import { useMediaQuery } from "@litespace/headless/mediaQuery";
 import { useOnError } from "@/hooks/error";
+import Close2 from "@litespace/assets/Close2";
+import { ITopic } from "@litespace/types";
+import { UseQueryResult } from "@tanstack/react-query";
 
 const TopicSelection: React.FC = () => {
   const intl = useFormatMessage();
   const mq = useMediaQuery();
-  const allTopicsQuery = useTopics({});
   const { query: userTopicsQuery } = useUserTopics();
   const [showDialog, setShowDialog] = useState(false);
   const toast = useToast();
   const invalidate = useInvalidateQuery();
-
-  const topicOptions = useMemo(() => {
-    if (!allTopicsQuery.query.data?.list) return [];
-    return allTopicsQuery.query.data.list.map((topic) => ({
-      id: topic.id,
-      label: topic.name.ar,
-    }));
-  }, [allTopicsQuery]);
+  const [selectedTopics, setSelectedTopics] = useState<
+    ITopic.PopulatedUserTopic[]
+  >([]);
 
   const userTopicIds = useMemo(() => {
     if (!userTopicsQuery.data) return [];
     return userTopicsQuery.data.map((topic) => topic.id);
   }, [userTopicsQuery.data]);
+
+  useEffect(() => {
+    if (!userTopicsQuery.data) return;
+    setSelectedTopics(userTopicsQuery.data);
+  }, [userTopicsQuery.data]);
+
+  const dataChanged = useMemo(() => {
+    return (
+      !userTopicIds ||
+      selectedTopics.some((topic) => !userTopicIds.includes(topic.id)) ||
+      userTopicIds.some((id) => !selectedTopics.map((s) => s.id).includes(id))
+    );
+  }, [userTopicIds, selectedTopics]);
+
+  const onRemoveTopic = useCallback((id: number) => {
+    setSelectedTopics((prev) => prev.filter((topic) => topic.id !== id));
+  }, []);
 
   const onSuccess = useCallback(() => {
     invalidate([QueryKey.FindUserTopics]);
@@ -54,8 +68,8 @@ const TopicSelection: React.FC = () => {
   });
 
   const updateTopics = useUpdateUserTopics({
-    onSuccess: onSuccess,
-    onError: onError,
+    onSuccess,
+    onError,
   });
 
   const confirm = useCallback(
@@ -76,8 +90,147 @@ const TopicSelection: React.FC = () => {
     [updateTopics, userTopicIds]
   );
 
+  if (userTopicsQuery.isPending)
+    return (
+      <TopicSelectionTemplate
+        confirm={confirm}
+        isUpdating={updateTopics.isPending}
+        setShowDialog={setShowDialog}
+        showDialog={showDialog}
+        userTopicsQuery={userTopicsQuery}
+        userTopicIds={userTopicIds}
+      >
+        <Loading
+          size="small"
+          text={intl("student-settings.topics.selection-dialog.loading")}
+        />
+      </TopicSelectionTemplate>
+    );
+
+  if (userTopicsQuery.isError && !userTopicsQuery.isPending)
+    return (
+      <TopicSelectionTemplate
+        confirm={confirm}
+        isUpdating={updateTopics.isPending}
+        setShowDialog={setShowDialog}
+        showDialog={showDialog}
+        userTopicsQuery={userTopicsQuery}
+        userTopicIds={userTopicIds}
+      >
+        <div className="flex justify-center">
+          <LoadingError
+            size="small"
+            retry={userTopicsQuery.refetch}
+            error={intl(
+              "student-settings.topics.selection-dialog.loading-error"
+            )}
+          />
+        </div>
+      </TopicSelectionTemplate>
+    );
+
+  if (isEmpty(selectedTopics))
+    return (
+      <TopicSelectionTemplate
+        confirm={confirm}
+        isUpdating={updateTopics.isPending}
+        setShowDialog={setShowDialog}
+        showDialog={showDialog}
+        userTopicsQuery={userTopicsQuery}
+        userTopicIds={userTopicIds}
+      >
+        <div className="flex justify-center w-full my-2">
+          <Animate>
+            <Button
+              size={mq.lg ? "large" : "medium"}
+              endIcon={
+                <AddCircle className="[&>*]:stroke-natural-50 w-4 h-4" />
+              }
+              onClick={() => setShowDialog(true)}
+            >
+              <Typography
+                tag="span"
+                className="text-tiny md:text-caption lg:text-body font-medium"
+              >
+                {intl("student-settings.add-topics-button.label")}
+              </Typography>
+            </Button>
+          </Animate>
+        </div>
+      </TopicSelectionTemplate>
+    );
+
   return (
-    <div className="flex flex-col gap-4 md:gap-6">
+    <TopicSelectionTemplate
+      confirm={confirm}
+      isUpdating={updateTopics.isPending}
+      setShowDialog={setShowDialog}
+      showDialog={showDialog}
+      userTopicsQuery={userTopicsQuery}
+      userTopicIds={userTopicIds}
+    >
+      <div className="grow md:grow-0 flex flex-col justify-between md:justify-start">
+        <div className="flex flex-row flex-wrap w-full gap-2 md:gap-y-4">
+          {selectedTopics.map((topic) => (
+            <Animate key={topic.id}>
+              <Typography
+                tag="span"
+                className="bg-brand-200 border border-brand-700 items-center flex text-brand-700 p-3 md:px-3 md:py-2 rounded-[24px] text-caption font-normal"
+              >
+                <button
+                  onClick={() => onRemoveTopic(topic.id)}
+                  className="cursor-pointer hidden md:block"
+                >
+                  <Close2 className="w-6 h-6 ml-[2px] [&>*]:fill-brand-700" />
+                </button>
+                {topic.name.ar}
+              </Typography>
+            </Animate>
+          ))}
+        </div>
+        <Button
+          size="large"
+          disabled={updateTopics.isPending || !dataChanged}
+          onClick={() => confirm(selectedTopics.map((s) => s.id))}
+          className="md:mt-10 mb-4 md:mb-0 mr-auto md:mr-0"
+        >
+          {intl("shared-settings.save")}
+        </Button>
+      </div>
+    </TopicSelectionTemplate>
+  );
+};
+
+const TopicSelectionTemplate = ({
+  userTopicIds,
+  setShowDialog,
+  showDialog,
+  confirm,
+  userTopicsQuery,
+  isUpdating,
+
+  children,
+}: {
+  userTopicIds: number[];
+  setShowDialog: (val: boolean) => void;
+  showDialog: boolean;
+  confirm: (topicIds: number[]) => void;
+  userTopicsQuery: UseQueryResult<ITopic.FindUserTopicsApiResponse, Error>;
+  isUpdating: boolean;
+  children: React.ReactNode;
+}) => {
+  const intl = useFormatMessage();
+  const allTopicsQuery = useTopics({});
+  const topicOptions = useMemo(() => {
+    if (!allTopicsQuery.query.data?.list) return [];
+    return allTopicsQuery.query.data.list.map((topic) => ({
+      id: topic.id,
+      label: topic.name.ar,
+    }));
+  }, [allTopicsQuery]);
+
+  return (
+    <div className="flex flex-col gap-4 grow">
       <div className="flex items-center justify-between">
         <Typography
           tag="h2"
@@ -87,112 +240,44 @@ const TopicSelection: React.FC = () => {
         </Typography>
 
         {!isEmpty(userTopicIds) ? (
-          <button
-            type="button"
-            onClick={() => {
-              setShowDialog(true);
-            }}
-          >
+          <button type="button" onClick={() => setShowDialog(true)}>
             <Typography
               tag="span"
-              className="flex text-brand-700 text-caption font-semibold"
+              className="flex text-natural-700 items-center text-body font-medium"
             >
+              <Edit className="w-4 h-4 ml-2 [&>*]:stroke-natural-700" />
               {intl("labels.update")}
-              <Edit className="w-6 h-6 mr-2 [&>*]:stroke-brand-700" />
             </Typography>
           </button>
         ) : null}
       </div>
 
-      <div className="flex flex-col gap-2">
-        <Typography
-          tag="span"
-          className="text-natural-950 hidden text-subtitle-2 font-normal"
-        >
-          {intl("student-settings.edit.personal.topics")}
-        </Typography>
+      {children}
 
-        {userTopicsQuery.isPending ? (
-          <div className="mt-6 sm:mt-8">
-            <Loading
-              size="small"
-              text={intl("student-settings.topics.selection-dialog.loading")}
-            />
-          </div>
-        ) : null}
-
-        {userTopicsQuery.isError && !userTopicsQuery.isPending ? (
-          <div className="mt-6 sm:mt-8">
-            <LoadingError
-              size="small"
-              retry={userTopicsQuery.refetch}
-              error={intl(
-                "student-settings.topics.selection-dialog.loading-error"
-              )}
-            />
-          </div>
-        ) : null}
-
-        {!userTopicsQuery.isPending && !userTopicsQuery.isError ? (
-          <div className="flex flex-row flex-wrap w-full gap-2">
-            {userTopicsQuery.data && !isEmpty(userTopicsQuery.data) ? (
-              userTopicsQuery.data.map((topic) => (
-                <div className="my-3" key={topic.id}>
-                  <Animate>
-                    <Typography
-                      tag="span"
-                      className="bg-brand-700 text-natural-50 px-4 py-3 rounded-[24px] text-caption font-normal"
-                    >
-                      {topic.name.ar}
-                    </Typography>
-                  </Animate>
-                </div>
-              ))
-            ) : (
-              <div className="flex justify-center w-full my-3">
-                <Animate>
-                  <Button
-                    size={mq.lg ? "large" : "medium"}
-                    endIcon={
-                      <AddCircle className="[&>*]:stroke-natural-50 w-4 h-4" />
-                    }
-                    onClick={() => setShowDialog(true)}
-                  >
-                    <Typography
-                      tag="span"
-                      className="text-tiny md:text-caption lg:text-body font-medium"
-                    >
-                      {intl("student-settings.add-topics-button.label")}
-                    </Typography>
-                  </Button>
-                </Animate>
-              </div>
-            )}
-          </div>
-        ) : null}
-      </div>
-
-      <TopicSelectionDialog
-        title={intl("student-settings.topics.selection-dialog.title")}
-        description={intl(
-          "student-settings.topics.selection-dialog.description"
-        )}
-        topics={topicOptions}
-        initialTopics={userTopicIds}
-        opened={showDialog}
-        retry={() => {
-          if (allTopicsQuery.query.isError) allTopicsQuery.query.refetch();
-          if (userTopicsQuery.isError) userTopicsQuery.refetch();
-        }}
-        confirming={updateTopics.isPending}
-        loading={userTopicsQuery.isPending || allTopicsQuery.query.isPending}
-        error={userTopicsQuery.isError || allTopicsQuery.query.isError}
-        close={() => {
-          setShowDialog(false);
-        }}
-        confirm={confirm}
-      />
+      {showDialog ? (
+        <TopicSelectionDialog
+          title={intl("student-settings.topics.selection-dialog.title")}
+          description={intl(
+            "student-settings.topics.selection-dialog.description"
+          )}
+          topics={topicOptions}
+          initialTopics={userTopicIds}
+          opened={showDialog}
+          retry={() => {
+            if (allTopicsQuery.query.isError) allTopicsQuery.query.refetch();
+            if (userTopicsQuery.isError) userTopicsQuery.refetch();
+          }}
+          confirming={isUpdating}
+          loading={userTopicsQuery.isPending || allTopicsQuery.query.isPending}
+          error={userTopicsQuery.isError || allTopicsQuery.query.isError}
+          close={() => {
+            setShowDialog(false);
+          }}
+          confirm={confirm}
+        />
+      ) : null}
     </div>
   );
 };
+
 export default TopicSelection;
