+++ apps/web/src/components/StudentSettings/UpdatePassword.tsx
@@ -84,8 +84,11 @@ export const UpdatePassword: React.FC<{ id: number }> = ({ id }) => {
   const mutation = useUpdateUser({ onSuccess, onError });
 
   return (
-    <div className="max-w-[400px]">
-      <form onSubmit={form.onSubmit} className="w-full flex flex-col gap-4">
+    <div className="max-w-[400px] grow md:grow-0 flex flex-col">
+      <form
+        onSubmit={form.onSubmit}
+        className="w-full flex flex-col grow gap-2 md:gap-4"
+      >
         <Password
           required
           name="current"
@@ -128,7 +131,7 @@ export const UpdatePassword: React.FC<{ id: number }> = ({ id }) => {
         size="large"
         disabled={mutation.isPending}
         onClick={form.submit}
-        className="mt-10"
+        className="md:mt-10 mb-4 md:mb-0 mr-auto md:mr-0"
       >
         {intl("shared-settings.save")}
       </Button>
