+++ packages/ui/src/locales/ar-eg.json
@@ -588,8 +588,6 @@
   "student-settings.topics.selection-dialog.loading-error": " عذرًا، حدث خطأ أثناء تحميل المفضلات. برجاء المحاولة مرة أخرى",
   "student-settings.topics.selection-dialog.update-error": " عذرًا، حدث خطأ أثناء تحديث المفضلات. برجاء المحاولة مرة أخرى",
   "student-settings.profile.title": "الإعدادات",
-  "student-settings.upload.image.label": "قم برفع صورة جديدة",
-  "student-settings.upload.image.label-sm": "تغيير الصورة",
   "student-settings.upload.image.change": "تغيير الصورة",
   "student-settings.upload.image.clear": "حذف الصورة",
   "student-settings.upload.image.clear.error": "حدث خطأ أثناء حذف الصورة برجاء المحاولة مرة أخرى!",
@@ -598,28 +596,12 @@
   "student-settings.verify-phone.title": "تأكيد رقم الهاتف",
   "student-settings.verify-phone.description": "أكد رقم هاتفك لتعزيز أمان حسابك، وتفعيل الإشعارات، واسترجاع الوصول إليه في حال فقدان كلمة المرور.",
   "student-settings.upload.image.desc": "يوصى بصيغة JPG أو PNG وGIF بحجم 480 × 480 بكسل على الأقل.",
-  "student-settings.edit.personal.title": "تعديل الملف الشخصي",
-  "student-settings.edit.personal.name": "الأسم",
-  "student-settings.edit.personal.name.placeholder": "قم بكتابة اسمك",
-  "student-settings.edit.personal.email": "البريد الإلكتروني",
-  "student-settings.edit.personal.email.placeholder": "قم بكتابة البريد الإلكتروني",
-  "student-settings.edit.personal.phone-number": "رقم الهاتف",
-  "student-settings.edit.personal.phone-number.placeholder": "قم بكتابة رقم هاتفك",
-  "student-settings.edit.personal.city": "المدينة",
   "student-settings.edit.personal.topics.title": "إضافة المفضلات",
   "student-settings.edit.personal.topics": "قم باختيار المجالات التي تريد التحدث عنها",
   "student-settings.edit.notification.label": "طريقة تلقي الإشعارات",
   "student-settings.edit.notification.placeholder": "واتساب/تيليجرام",
   "student-settings.edit.notification.whatsapp": "واتساب",
   "student-settings.edit.notification.telegram": "تيليجرام",
-  "student-settings.loading": "برجاء الإنتظار... جاري تحميل الإعدادات!",
-  "student-settings.error": " عذرًا، حدث خطأ أثناء تحميل  الإعدادات، برجاء المحاولة مرة أخرى",
-  "student-settings.confirm-email.title": "تأكيد الحساب الخاص بك",
-  "student-settings.confirm-email.desc": "أكد حسابك حتي تتمكن من حجز الحصص او التحدث مع المعلمين، تفقد بريدك الإلكتروني لتجد رابط التاكيد.",
-  "student-settings.confirm-email.confirm": "تأكيد الحساب",
-  "student-settings.confirm-email.error": "عذرًا، حدث خطأ أثناء إرسال رابط التأكيد إلى بريدك الإلكتروني.",
-  "student-settings.confirm-email.success.title": "تم إرسال رابط التأكيد إلى بريدك الإلكتروني بنجاح.",
-  "student-settings.confirm-email.success.desc": "يرجى التحقق من صندوق الوارد الخاص بك واتباع التعليمات لتأكيد حسابك",
   "shared-settings.edit.personal.city.placeholder": "الجيزة",
   "shared-settings.save": "حفظ التغيرات",
   "shared-settings.update.error": "لديك مشكلة في تحديث البيانات",
@@ -634,7 +616,6 @@
   "shared-settings.notifications.messages.description": "سوف يتم اشعارك بجميع الرسائل الواردة",
   "shared-settings.notifications.coming-soon.title": "قريبا 🎉",
   "shared-settings.notifications.coming-soon.description": "بإمكانك التحكم في الاشعارات بمجرد ان يتم تفعيل هذه الخاصية.",
-  "shared-settings.edit.password.title": "تغيير كلمة المرور",
   "shared-settings.edit.password.current": "كلمة المرور الحالية",
   "shared-settings.edit.password.new": "كلمة المرور الجديدة",
   "shared-settings.edit.password.confirm": "تاكيد كلمة المرور الجديدة",
